requirements.txt
asgiref==3.7.2
certifi==2023.7.22
distlib==0.3.7
Django==4.2.6
django-cors-headers==4.3.0
django-environ==0.11.2
djangorestframework==3.14.0
filelock==3.12.4
pbr==5.11.1
platformdirs==3.11.0
psycopg==3.1.12
pytz==2023.3.post1
sqlparse==0.4.4
stevedore==5.1.0
typing_extensions==4.8.0
virtualenv==20.24.5
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
